Added a test class for the individual class. Added tests for constructors and set methods

+++ MutationRatesGeneticAlgorithm/src/geneticAlgorithm/Individual.java
@@ -134,7 +134,7 @@ public class Individual {
 	/*
 	 * sets the mutation rate
 	 */
-	public void setMuationRate(int mutationRate) {
+	public void setMuationRate(double mutationRate) {
 		this.mutationRate = mutationRate;
 	}
 }
